fix misspelling of BlankCPlusPlusLanguageTemplate
